reorganize and add to the lib directory
