+++ src/lib/schema/normalize.test.js
