+++ src/lib/schema/schema-parser.js
@@ -1,9 +1,9 @@
 import $RefParser from 'json-schema-ref-parser';
 import { getAttributes, getRelationships } from './normalize';
-import { request } from './request';
-import { extract } from '../utils';
-import Document from './document';
-import { compileJsonApiUrl, parseJsonApiUrl } from './url';
+import { request } from '../utils/request';
+import { extract } from '../utils/utils';
+import Document from '../jsonapi-objects/document';
+import { compileJsonApiUrl, parseJsonApiUrl } from '../url/url';
 
 export default class SchemaParser {
   constructor() {
