+++ src/schema-ui.js
@@ -3,7 +3,7 @@ import React, { useContext } from 'react';
 import SchemaAttributes from './schemaAttributes';
 import SchemaRelationships from './schemaRelationships';
 
-import { checkIncludesPath } from './utils';
+import { checkIncludesPath } from './lib/utils/utils';
 import { LocationContext } from './contexts/location';
 import { SchemaContext } from './contexts/schema';
